+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #
 * Add `Container::restore`, `Container::resize`, `Container::export`
 * Add support for `!=` filters like `label!=key=val` or `label!=key`
+* Add missing filter function `NetworkPruneOptsBuilder::filters`
 
 # 0.5.0
 * Fix some generated models